third_party/WebKit/LayoutTests/fast/borders/border-image-slice-parsing-support.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_true(CSS.supports('border-image-slice', '1.2'), 'accepts non-integer numbers');
  assert_true(CSS.supports('border-image-slice', '1 2'), 'accepts integers');
  assert_true(CSS.supports('border-image-slice', '1.2 12'), 'accepts both integers and non-integers');
  assert_false(CSS.supports('border-image-slice', '1 -2'), 'rejects negative integers');
  assert_false(CSS.supports('border-image-slice', '1 -1.2'), 'rejects negative non-integers');
}, 'border-image-slice parsing support');
</script>
